requirements.txt
Django>=3.2.8,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9.1,<2.10.0

flake8>=4.0.1,<4.2.0